requirements.txt
python-multipart
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
qrcode
FastAPI-SQLAlchemy
psycopg2
pandas
Image